Move `oxql` from `oxdb` to mainline `omdb`
Why?:

Concerning [network observability work](https://github.com/orgs/oxidecomputer/projects/55/views/1?filterQuery=&pane=issue&itemId=68336554),
this makes the [`oxql`](https://rfd.shared.oxide.computer/rfd/0463) interactive query repl accessible via omdb, as we start to give users and
ourselves the ability to query timeseries and metrics more easily. Additionally, in the "now", this aids in debugging through our metrics
set and makes it available, via omdb, throughout our ecosystem/a4x2.

Includes:
  * Moves `oxql` and `sql` shells (from oxdb) into the oximeter_db src for use by both omdb (mainline tool) and oxdb.
    * simplifies the oxdb bin to leverage the new `shells` module
  * If no URL is given to `omdb oxql`, it will leverage internal DNS.
  * Update the oximeter omdb call (for listing producers) to leverage internal DNS if no URL is given.
  * Update command/output tests/generations and collector specific tests for list producers.

Notes:
  * The oxql client still expects an socket address as liked it typed
    specifically v.s. a String. Instead, upon running the `omdb oxql` command,
    we take in a URL String and parse it into the socket address directly.
