+++ dev-tools/omdb/src/bin/omdb/main.rs
@@ -50,6 +50,7 @@ mod helpers;
 mod mgs;
 mod nexus;
 mod oximeter;
+mod oxql;
 mod sled_agent;
 
 #[tokio::main]
@@ -66,7 +67,8 @@ async fn main() -> Result<(), anyhow::Error> {
         OmdbCommands::Db(db) => db.run_cmd(&args, &log).await,
         OmdbCommands::Mgs(mgs) => mgs.run_cmd(&args, &log).await,
         OmdbCommands::Nexus(nexus) => nexus.run_cmd(&args, &log).await,
-        OmdbCommands::Oximeter(oximeter) => oximeter.run_cmd(&log).await,
+        OmdbCommands::Oximeter(oximeter) => oximeter.run_cmd(&args, &log).await,
+        OmdbCommands::Oxql(oxql) => oxql.run_cmd(&args, &log).await,
         OmdbCommands::SledAgent(sled) => sled.run_cmd(&args, &log).await,
         OmdbCommands::CrucibleAgent(crucible) => crucible.run_cmd(&args).await,
     }
@@ -269,6 +271,8 @@ enum OmdbCommands {
     Nexus(nexus::NexusArgs),
     /// Query oximeter collector state
     Oximeter(oximeter::OximeterArgs),
+    /// Enter the Oximeter Query Language shell for interactive querying.
+    Oxql(oxql::OxqlArgs),
     /// Debug a specific Sled
     SledAgent(sled_agent::SledAgentArgs),
 }
