+++ dev-tools/omdb/tests/env.out
@@ -433,3 +433,29 @@ note: using database URL postgresql://root@[::1]:REDACTED_PORT/omicron?sslmode=d
 note: database schema version matches expected (<redacted database version>)
 note: listing all commissioned sleds (use -F to filter, e.g. -F in-service)
 =============================================
+EXECUTING COMMAND: omdb ["oximeter", "--oximeter-url", "junk", "list-producers"]
+termination: Exited(1)
+---------------------------------------------
+stdout:
+---------------------------------------------
+stderr:
+note: using Oximeter URL junk
+Error: failed to fetch collector info
+
+Caused by:
+    0: Communication Error: builder error: relative URL without a base
+    1: builder error: relative URL without a base
+    2: relative URL without a base
+=============================================
+EXECUTING COMMAND: omdb ["oxql", "--metrics-db-url", "junk"]
+termination: Exited(2)
+---------------------------------------------
+stdout:
+---------------------------------------------
+stderr:
+error: unexpected argument '--metrics-db-url' found
+
+Usage: omdb oxql [OPTIONS]
+
+For more information, try '--help'.
+=============================================
