+++ oximeter/db/src/client/mod.rs
@@ -7,7 +7,6 @@
 // Copyright 2024 Oxide Computer Company
 
 pub(crate) mod dbwrite;
-#[cfg(any(feature = "oxql", test))]
 pub(crate) mod oxql;
 pub(crate) mod query_summary;
 #[cfg(any(feature = "sql", test))]
