+++ oximeter/db/src/oxql/query/mod.rs
@@ -25,63 +25,6 @@ use chrono::DateTime;
 use chrono::Utc;
 use std::time::Duration;
 
-/// Special identifiers for column names or other widely-used values.
-pub mod special_idents {
-    use oximeter::DatumType;
-
-    macro_rules! gen_marker {
-        ($p:expr, $field:expr) => {
-            concat!("p", $p, "_", $field)
-        };
-    }
-
-    pub const TIMESTAMP: &str = "timestamp";
-    pub const START_TIME: &str = "start_time";
-    pub const DATUM: &str = "datum";
-    pub const BINS: &str = "bins";
-    pub const COUNTS: &str = "counts";
-    pub const MIN: &str = "min";
-    pub const MAX: &str = "max";
-    pub const SUM_OF_SAMPLES: &str = "sum_of_samples";
-    pub const SQUARED_MEAN: &str = "squared_mean";
-    pub const DATETIME64: &str = "DateTime64";
-    pub const ARRAYU64: &str = "Array[u64]";
-    pub const ARRAYFLOAT64: &str = "Array[f64]";
-    pub const ARRAYINT64: &str = "Array[i64]";
-    pub const FLOAT64: &str = "f64";
-    pub const UINT64: &str = "u64";
-
-    pub const DISTRIBUTION_IDENTS: [&str; 15] = [
-        "bins",
-        "counts",
-        "min",
-        "max",
-        "sum_of_samples",
-        "squared_mean",
-        gen_marker!("50", "marker_heights"),
-        gen_marker!("50", "marker_positions"),
-        gen_marker!("50", "desired_marker_positions"),
-        gen_marker!("90", "marker_heights"),
-        gen_marker!("90", "marker_positions"),
-        gen_marker!("90", "desired_marker_positions"),
-        gen_marker!("99", "marker_heights"),
-        gen_marker!("99", "marker_positions"),
-        gen_marker!("99", "desired_marker_positions"),
-    ];
-
-    pub fn array_type_name_from_histogram_type(
-        type_: DatumType,
-    ) -> Option<String> {
-        if !type_.is_histogram() {
-            return None;
-        }
-        Some(format!(
-            "Array[{}]",
-            type_.to_string().strip_prefix("Histogram").unwrap().to_lowercase(),
-        ))
-    }
-}
-
 /// A parsed OxQL query.
 #[derive(Clone, Debug, PartialEq)]
 pub struct Query {
